refactor(feedback02): 과추상 정상화
